Split GHC.Hs.Instances for parallel compilation (#18254)
Split the monolithic GHC.Hs.Instances module (344 Data instances, 14MB
object file) into 5 parallel-compilable sub-modules:

  - GHC.Hs.Instances.Common      (phase-independent instances)
  - GHC.Hs.Instances.Transitions (LR types spanning multiple phases)
  - GHC.Hs.Instances.Parsed      (GhcPs instances)
  - GHC.Hs.Instances.Renamed     (GhcRn instances + HsLocalBinds deps)
  - GHC.Hs.Instances.Typechecked (GhcTc instances)

The wrapper module GHC.Hs.Instances re-exports all sub-modules for
backward compatibility.

Add stub object-code linker functions to GHC.Runtime.Interpreter that
panic when called (initObjLinker, lookupSymbol, loadDLL, etc.) for
no-interpreter builds. Guard Interpreter/Init.hs and
Config/Interpreter.hs with HAVE_INTERPRETER.
